Move sparql connector logic
