+++ packages/graph-explorer/src/connector/sparql/fetchBlankNodeNeighbors/blankNodeOneHopNeighborsTemplate.ts
