+++ packages/graph-explorer/src/connector/sparql/fetchSchema/classesWithCountsTemplates.ts
