+++ packages/graph-explorer/src/connector/sparql/fetchSchema/predicatesByClassTemplate.ts
