+++ packages/graph-explorer/src/connector/sparql/fetchSchema/predicatesWithCountsTemplate.ts
