+++ packages/graph-explorer/src/connector/sparql/fetchVertexCountsByType/classWithCountsTemplates.ts
