+++ packages/graph-explorer/src/connector/sparql/keywordSearch/helpers.ts
