+++ packages/graph-explorer/src/connector/sparql/sparqlExplorer.ts
@@ -2,19 +2,14 @@ import type {
   Criterion,
   Explorer,
   ExplorerRequestOptions,
-  KeywordSearchRequest,
-  KeywordSearchResponse,
-  NeighborsResponse,
 } from "../useGEFetchTypes";
 import { fetchDatabaseRequest } from "../fetchDatabaseRequest";
-import fetchBlankNodeNeighbors from "./queries/fetchBlankNodeNeighbors";
-import fetchClassCounts from "./queries/fetchClassCounts";
-import fetchNeighbors from "./queries/fetchNeighbors";
-import fetchNeighborsCount from "./queries/fetchNeighborsCount";
-import fetchSchema from "./queries/fetchSchema";
-import keywordSearch from "./queries/keywordSearch";
-import keywordSearchBlankNodesIdsTemplate from "./templates/keywordSearch/keywordSearchBlankNodesIdsTemplate";
-import oneHopNeighborsBlankNodesIdsTemplate from "./templates/oneHopNeighbors/oneHopNeighborsBlankNodesIdsTemplate";
+import fetchBlankNodeNeighbors from "./fetchBlankNodeNeighbors";
+import fetchClassCounts from "./fetchVertexCountsByType";
+import fetchNeighbors from "./fetchNeighbors";
+import fetchNeighborsCount from "./fetchNeighborsCount";
+import fetchSchema from "./fetchSchema";
+import keywordSearch from "./keywordSearch";
 import {
   BlankNodesMap,
   GraphSummary,
@@ -26,122 +21,9 @@ import { v4 } from "uuid";
 import { env, logger } from "@/utils";
 import { createLoggerFromConnection } from "@/core/connector";
 import { FeatureFlags } from "@/core";
-
-const replaceBlankNodeFromSearch = (
-  blankNodes: BlankNodesMap,
-  request: KeywordSearchRequest,
-  response: KeywordSearchResponse
-) => {
-  logger.log(
-    "[SPARQL Explorer] Replacing blank node from search with keywordSearchBlankNodesIdsTemplate"
-  );
-  return response.vertices.map(vertex => {
-    if (!vertex.__isBlank) {
-      return vertex;
-    }
-
-    const bNode = blankNodes.get(vertex.id);
-
-    if (!bNode) {
-      blankNodes.set(vertex.id, {
-        id: vertex.id,
-        subQueryTemplate: keywordSearchBlankNodesIdsTemplate(request),
-        vertex,
-        neighborCounts: {
-          totalCount: 0,
-          counts: {},
-        },
-      });
-    }
-
-    return bNode?.vertex ?? vertex;
-  });
-};
-
-const replaceBlankNodeFromNeighbors = (
-  blankNodes: BlankNodesMap,
-  request: SPARQLNeighborsRequest,
-  response: KeywordSearchResponse
-) => {
-  logger.log(
-    "[SPARQL Explorer] Replacing blank node from search with oneHopNeighborsBlankNodesIdsTemplate"
-  );
-  return response.vertices.map(vertex => {
-    if (!vertex.__isBlank) {
-      return vertex;
-    }
-
-    const bNode = blankNodes.get(vertex.id);
-    if (!bNode?.neighbors) {
-      blankNodes.set(vertex.id, {
-        id: vertex.id,
-        subQueryTemplate: oneHopNeighborsBlankNodesIdsTemplate(request),
-        vertex,
-        neighborCounts: {
-          totalCount: 0,
-          counts: {},
-        },
-      });
-    }
-
-    return bNode?.vertex ?? vertex;
-  });
-};
-/**
- * This mock request takes into account the request filtering
- * to narrow the neighbors results of the given blank node.
- */
-const storedBlankNodeNeighborsRequest = (
-  blankNodes: BlankNodesMap,
-  req: SPARQLNeighborsRequest
-) => {
-  return new Promise<NeighborsResponse>(resolve => {
-    const bNode = blankNodes.get(req.resourceURI);
-    if (!bNode?.neighbors) {
-      resolve({
-        vertices: [],
-        edges: [],
-        scalars: [],
-      });
-      return;
-    }
-
-    const filteredVertices = bNode.neighbors.vertices.filter(vertex => {
-      if (!req.subjectClasses && !req.filterCriteria?.length) {
-        return true;
-      }
-
-      if (!req.subjectClasses?.includes(vertex.type)) {
-        return false;
-      }
-
-      if (!req.filterCriteria?.length) {
-        return true;
-      }
-
-      for (const criterion of req.filterCriteria) {
-        const attrVal = vertex.attributes[criterion.predicate];
-        if (attrVal == null) {
-          return false;
-        }
-        if (!String(attrVal).match(new RegExp(criterion.object, "gi"))) {
-          return false;
-        }
-      }
-
-      return true;
-    });
-
-    resolve({
-      vertices: filteredVertices.slice(
-        req.offset ?? 0,
-        req.limit ? req.limit + (req.offset ?? 0) : undefined
-      ),
-      edges: bNode.neighbors.edges,
-      scalars: [],
-    });
-  });
-};
+import { replaceBlankNodeFromNeighbors } from "./fetchNeighbors/replaceBlankNodeFromNeighbors";
+import { storedBlankNodeNeighborsRequest } from "./fetchNeighbors/storedBlankNodeNeighborsRequest";
+import { replaceBlankNodeFromSearch } from "./keywordSearch/replaceBlankNodeFromSearch";
 
 function _sparqlFetch(
   connection: ConnectionConfig,
